# app/services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for KienlongBank Treasury Management System — Phase 1",
        "title": "Treasury Management System API",
        "termsOfService": "https://kienlongbank.com/terms",
        "contact": {
            "name": "KienlongBank IT Department",
            "email": "it@kienlongbank.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://kienlongbank.com/license"
        },
        "version": "1.0"
    },
    "host": "localhost:34000",
    "basePath": "/api/v1",
    "paths": {
...
        "/fx/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Phê duyệt (APPROVE) hoặc từ chối (REJECT) giao dịch ngoại tệ. Yêu cầu quyền FX_DEAL.APPROVE.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FX"
                ],
                "summary": "Phê duyệt / Từ chối giao dịch FX",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FX Deal ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Approval payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.ApprovalRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.ResponseMeta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/fx/{id}/clone": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tạo bản sao của giao dịch ngoại tệ hiện có. Giao dịch mới ở trạng thái OPEN.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FX"
                ],
                "summary": "Sao chép giao dịch FX",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FX Deal ID to clone (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.FxDealResponse"
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.ResponseMeta"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_kienlongbank_treasury-api_pkg_dto.APIError": {
            "description": "Standard error body with optional field-level validation errors.",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_kienlongbank_treasury-api_pkg_dto.APIResponse": {
            "description": "Standard API response envelope.",
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.ResponseMeta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_kienlongbank_treasury-api_pkg_dto.ApprovalRequest": {
            "description": "Request body for deal approval or rejection.",
            "type": "object",
            "required": [
                "action",
                "version"
            ],
            "properties": {
                "action": {
                    "type": "string",
                    "enum": [
                        "APPROVE",
                        "REJECT"
                    ]
                },
                "comment": {
                    "type": "string",
                    "maxLength": 2000
                },
                "version": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "github_com_kienlongbank_treasury-api_pkg_dto.ResponseMeta": {
            "description": "Metadata about the API response including request tracing and pagination.",
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/github_com_kienlongbank_treasury-api_pkg_dto.PaginationMeta"
                },
                "request_id": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token (JWT)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}